:building_construction: :truck: Implemented the `Fractal` file structure
