+++ app/screens/CounterPage.js
@@ -1,7 +1,7 @@
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import Counter from '../components/Counter';
-import * as CounterActions from '../actions/counter';
+import { CounterActions } from '../modules/counter';
 
 function mapStateToProps(state) {
   return {
