+++ app/store/configureStore.prod.js
@@ -3,8 +3,8 @@ import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import { createHashHistory } from 'history';
 import { routerMiddleware } from 'react-router-redux';
-import rootReducer from '../reducers';
-import type { counterStateType } from '../reducers/counter';
+import rootReducer from '../modules';
+import type { counterStateType } from '../modules/counter';
 
 const history = createHashHistory();
 const router = routerMiddleware(history);
